Added a chatGPT creted code based on my prompt. / I want to write a python application that can interact with the HP OneView API and get information about the server hardware in a 3 frame configuration. The API reference guide can be found here. https://support.hpe.com/docs/display/public/dp00003325en_us/index.html#rest/server-hardware
This application will need to do the following, it will need to ask for a username and password to authentication to the oneview instance to get a sessionid, this sessionid is our auth token for all subsequent requests to the api. The API version being used is 3800, once we have an auth token, we need to gather the following information from each server hardware device in the instance. CPU, Model type, Model number, Processor type, Memory as well as stick sizing E.G. 16 or 32 GB dimms. Storage controller, number of HDDs and storage capacity. Then it needs to place this information into a spreadsheet organized by those items